# optimize child views in template