--- a/lib/viewing/RenderableView.js
+++ b/lib/viewing/RenderableView.js
@@ -5,6 +5,7 @@ var HasChildViews = require("../viewing/HasChildViews");
 var ReattachableView = require("../viewing/ReattachableView");
 var StringTemplateAdapter = require("../templating/StringTemplateAdapter");
 var TemplateAdapter = require("../templating/TemplateAdapter");
+var Environment = require("../environment/Environment");
 var noop = require("../util/noop");
 
 var VIEW_PLACEHOLDER_PREFIX = "b_-_t";
@@ -67,8 +68,15 @@ var RenderableView = _.extendOwn({},
 
                 var placeholderId = VIEW_PLACEHOLDER_PREFIX + i;
 
-                this.viewPlaceholders[identifier] = viewPlaceholder(placeholderId);
-                viewRelationship.andReplace("#" + placeholderId);
+                if (Environment.isClient()) {
+                    this.viewPlaceholders[identifier] = viewPlaceholder(placeholderId);
+                    viewRelationship.andReplace("#" + placeholderId);
+                }
+
+                if (Environment.isServer()) {
+                    this.viewPlaceholders[identifier] = asString(viewRelationship);
+                }
+
             });
         },
 
@@ -190,6 +198,14 @@ function modelToJSON(model) {
 
 function viewPlaceholder(placeholderId) {
     return "<div id='" + placeholderId + "'></div>";
+}
+
+function asString(viewRelationship) {
+    return {
+        toString: function() {
+            return viewRelationship.asString();
+        }
+    };
 }
 
 module.exports = RenderableView;