--- a/lib/viewing/ViewRelationship.js
+++ b/lib/viewing/ViewRelationship.js
@@ -129,6 +129,12 @@ ViewRelationship.prototype = {
         this.enterDOM();
     },
 
+    asString: function() {
+        this.instantiateChildView();
+
+        return this.childView.render().el.outerHTML;
+    },
+
     close: function() {
         if (this.childView && this.childView.close) {
             this.childView.close();