--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "brisket",
-  "version": "0.46.0",
+  "version": "0.46.1",
   "description": "Framework for building single page web apps using isomorphic Javascript",
   "author": [
     "Wayne Warner",