before this change:

```js
[00:09:06] Running suite Rendering child views in template...
[00:09:11]    child views in template x 3,714 ops/sec ±1.06% (92 runs sampled)
[00:09:17]    child views by replacing a node via parent view x 3,710 ops/sec ±1.77% (92 runs sampled)
[00:09:22]    child views by appending them via parent view x 5,837 ops/sec ±2.09% (89 runs sampled)
[00:09:22] Fastest test is child views by appending them via parent view at 1.57x faster than child views by replacing a node via parent view and child views in template
```

after the change:

```bash
[00:08:10] Running suite Rendering child views in template...
[00:08:15]    child views in template x 5,637 ops/sec ±0.94% (93 runs sampled)
[00:08:21]    child views by replacing a node via parent view x 3,376 ops/sec ±2.84% (86 runs sampled)
[00:08:27]    child views by appending them via parent view x 5,898 ops/sec ±2.04% (90 runs sampled)
[00:08:27] Fastest test is child views by appending them via parent view at 1.05x faster than child views in template
```