@wawjr3d 2016-02-04
if you do something like the following, the change in this PR changes behavior:

```js
const MyView = Brisket.View.extend({

  template({ views }) {
    return `${views.child1}`;
  },

  beforeRender() {
    this.view = new Brisket.View();
    this.createChildView('child1', this.view);
  },

  afterRender() {
    this.view.$el.addClass('expect-class-to-be-there-on-server');
  }
});
```

when this View renders, the commands to this.view won't have any effect on the server side render. the commands won't take effect until the view reattaches in the browser. 

i think coding a child view like this is a strange edge case. I would sacrifice it for the speed boost.